+++ tests/roots/test-ext-agentskills-nopy/conf.py
@@ -1,0 +1,5 @@
+extensions = ['sphinx.ext.agentskills']
+
+project = 'NoPI'
+version = '0.1'
+release = '0.1'