+++ tests/roots/test-ext-agentskills/conf.py
@@ -1,0 +1,5 @@
+extensions = ['sphinx.ext.agentskills']
+
+project = 'TestProject'
+version = '1.2'
+release = '1.2.3'